color: clamp Pixel16::operator* in float to avoid out-of-range cast (P2-7)
(int)(r * s) is UB when r*s exceeds INT_MAX for a large gain s; the
int clamp ran too late. Clamp against [0, 65535] in float before the
cast so the operator is safe for any scale.

+++ core/color.h
@@ -86,9 +86,12 @@ struct Pixel16 {
   }
 
   Pixel16 operator*(float s) const {
-    return Pixel16((uint16_t)std::clamp((int)(r * s), 0, 65535),
-                   (uint16_t)std::clamp((int)(g * s), 0, 65535),
-                   (uint16_t)std::clamp((int)(b * s), 0, 65535));
+    // Clamp in float before the cast: r*s can exceed INT_MAX for large s,
+    // and float->int conversion is UB out of range (would bite before an
+    // int clamp could fire).
+    return Pixel16((uint16_t)std::clamp(r * s, 0.0f, 65535.0f),
+                   (uint16_t)std::clamp(g * s, 0.0f, 65535.0f),
+                   (uint16_t)std::clamp(b * s, 0.0f, 65535.0f));
   }
 
   // Lerp (Linear interpolation 16-bit)
